requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
ibm-cloud-sdk-core==3.13.2
ibm-watson==5.3.0
idna==3.3
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
websocket-client==1.1.0
